Update sweepers for new organization.
