+++ digitalocean/droplet/resource_droplet_test.go
@@ -1,58 +1,19 @@
 package droplet_test
 
 import (
-	"context"
 	"fmt"
-	"log"
 	"regexp"
 	"strings"
 	"testing"
 
 	"github.com/digitalocean/godo"
 	"github.com/digitalocean/terraform-provider-digitalocean/digitalocean/acceptance"
-	"github.com/digitalocean/terraform-provider-digitalocean/digitalocean/config"
 	"github.com/digitalocean/terraform-provider-digitalocean/digitalocean/droplet"
 	"github.com/digitalocean/terraform-provider-digitalocean/digitalocean/util"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
-
-func init() {
-	resource.AddTestSweepers("digitalocean_droplet", &resource.Sweeper{
-		Name: "digitalocean_droplet",
-		F:    testSweepDroplets,
-	})
-
-}
-
-func testSweepDroplets(region string) error {
-	meta, err := acceptance.SharedConfigForRegion(region)
-	if err != nil {
-		return err
-	}
-
-	client := meta.(*config.CombinedConfig).GodoClient()
-
-	opt := &godo.ListOptions{PerPage: 200}
-	droplets, _, err := client.Droplets.List(context.Background(), opt)
-	if err != nil {
-		return err
-	}
-	log.Printf("[DEBUG] Found %d droplets to sweep", len(droplets))
-
-	for _, d := range droplets {
-		if strings.HasPrefix(d.Name, "foo-") || strings.HasPrefix(d.Name, "bar-") || strings.HasPrefix(d.Name, "baz-") || strings.HasPrefix(d.Name, "tf-acc-test-") || strings.HasPrefix(d.Name, "foobar-") {
-			log.Printf("Destroying Droplet %s", d.Name)
-
-			if _, err := client.Droplets.Delete(context.Background(), d.ID); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
 
 func TestAccDigitalOceanDroplet_Basic(t *testing.T) {
 	var droplet godo.Droplet
