+++ digitalocean/reservedip/resource_floating_ip_test.go
@@ -14,36 +14,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
-func init() {
-	resource.AddTestSweepers("digitalocean_floating_ip", &resource.Sweeper{
-		Name: "digitalocean_floating_ip",
-		F:    testSweepFloatingIps,
-	})
-
-}
-
-func testSweepFloatingIps(region string) error {
-	meta, err := acceptance.SharedConfigForRegion(region)
-	if err != nil {
-		return err
-	}
-
-	client := meta.(*config.CombinedConfig).GodoClient()
-
-	ips, _, err := client.FloatingIPs.List(context.Background(), nil)
-	if err != nil {
-		return err
-	}
-
-	for _, ip := range ips {
-		if _, err := client.FloatingIPs.Delete(context.Background(), ip.IP); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 func TestAccDigitalOceanFloatingIP_Region(t *testing.T) {
 	var floatingIP godo.FloatingIP
